fix locality issue in mocha unit tests, move sample helper functions for localStorage unit testing
